pyproject.toml
[tool.poetry]
name = "title-evaluator"
version = "0.1.0"
description = "A title evaluation agent using LangChain and LangGraph"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.1.0"
langgraph = "^0.0.10"
langchain-openai = "^0.0.2"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 